Create dmart_bill.py
The Primary objective of this “Generating d-mart bill using object oriented programming in 
python” is to save time of the customers and the company. It also helps us to keep records of the 
purchased items by the customer in a day.
This project helps us to identify the frauds. It also helps us to identify the profits and losses.